+++ data_fast_insights/_binary_dependence_model_data.py
@@ -3,13 +3,24 @@ from typing import Iterable, Optional
 from itertools import combinations
 import logging
 from collections import OrderedDict
-import json
+import enum
 
 import numpy as np
 import pandas as pd
 
+from ._decision_tree_multidim_binning import get_optimized_combination_splits, compile_splits_to_string_interval
+from .resources.const_names import ENTIRE_DATASET_SEGMENT_NAME
+
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
+
+
+class SegmentSource(enum.Enum):
+    PRIMARY_BINNING = 1
+    PARTIAL_COMBINATION = 2
+    COMBINATION_ON_EXISTING = 3
+    DECISION_TREE_BINNING = 4
+    ENTIRE_DATASET = 5
 
 
 class BinaryDependenceModelData:
@@ -21,6 +32,9 @@ class BinaryDependenceModelData:
         Attributes description is to be done.
     """
 
+    # TODO: use them dynamically as well
+    RESERVED_SUBSTRINGS = {'_AND_', '_IN_', '_NOT_IN_', '<=', '>', ENTIRE_DATASET_SEGMENT_NAME}
+
     # TODO: attributes description, since they are useful
     def __init__(self,
                  base_data: pd.DataFrame,
@@ -28,8 +42,10 @@ class BinaryDependenceModelData:
                  cat_cols: Optional[Iterable[str]] = None,
                  num_cols: Optional[Iterable[str]] = None,
                  y_type: Optional[str] = "quantile",
+                 y_quantile: Optional[float] = 0.5,
                  exclude_zero_var: Optional[bool] = True,
-                 **kwargs) -> None:
+                 inverse_goal: Optional[bool] = False,
+                 ) -> None:
         """ Initialize object that holds all information about features and target in its attributes.
             This object is supposed to be used further in the calculations of target analysis model.
 
@@ -61,10 +77,24 @@ class BinaryDependenceModelData:
                 - checks numeric features and excludes those having variance = 0
             In case you've passed a dataset with such features already excluded, you might set this to False
             for potential speed up
+
+        inverse_goal
+            EXPERIMENTAL
+            By default higher values of target are considered "better". Setting this to True
+            will make it so that higher values are considered to be worsening the target.
+            E.g. if target is handling time and you want to reduce it
         """
         if not isinstance(base_data, pd.DataFrame):
             raise TypeError('base_data argument must be a DataFrame object')
         self.base_data = base_data.copy()
+
+        # CHECK RESERVED SUBSTRINGS
+        # I don't know if allowing them will cause problems, but just in case we'll disallow them
+        for col in set(cat_cols).union(set(num_cols)).union({y_name}):
+            for reserve in self.RESERVED_SUBSTRINGS:
+                if reserve in col:
+                    raise ValueError(f'Reserved substring {reserve} was found in column name: {col}. '
+                                     f'Please rename the column.')
 
         # SET CATEGORICAL AND NUMERIC COLUMNS
         try:
@@ -89,7 +119,7 @@ class BinaryDependenceModelData:
         self.target_processing_attrs = dict()
         if y_type == 'quantile':
             self.target_processing_attrs['y_type'] = y_type
-            self.target_processing_attrs['y_quantile'] = kwargs.get('y_quantile', 0.5)
+            self.target_processing_attrs['y_quantile'] = y_quantile
         elif y_type in ('mean', 'binary'):
             self.target_processing_attrs['y_type'] = y_type
         else:
@@ -97,6 +127,7 @@ class BinaryDependenceModelData:
 
         # SET OTHER
         self.exclude_zero_var = exclude_zero_var
+        self.inverse_goal = inverse_goal
 
         # Data for converted features
         self.data = self.base_data[[self.y_name]].copy()
@@ -106,6 +137,8 @@ class BinaryDependenceModelData:
         self.bins = None
         self.y_binary_name = None
         self.is_data_converted = False
+
+        self.segment_sources: dict[str, SegmentSource] = dict()
 
         self._reset_binary_data()
 
@@ -139,7 +172,7 @@ class BinaryDependenceModelData:
                     exclude_cats.append(cat)
             for cat in exclude_cats:
                 self.cat_cols.remove(cat)
-                self.base_data = self.base_data.drop(cat, 1)
+                self.base_data = self.base_data.drop(cat, axis=1)
                 logging.warning(f"{cat} feature was removed before the analysis, because it has < 2 unique values")
 
             for num in self.num_cols:
@@ -147,7 +180,7 @@ class BinaryDependenceModelData:
                     exclude_nums.append(num)
             for num in exclude_nums:
                 self.num_cols.remove(num)
-                self.base_data = self.base_data.drop(num, 1)
+                self.base_data = self.base_data.drop(num, axis=1)
                 logging.warning(f"{num} feature was removed before the analysis, because it has zero variance")
 
     def _convert_types(self) -> None:
@@ -184,7 +217,13 @@ class BinaryDependenceModelData:
             return
 
         self.y_binary_name = 'is_' + self.y_name + '_lt_' + self.target_processing_attrs['y_type']
-        self.data[self.y_binary_name] = self.base_data[self.y_name] < self.y_pivot
+
+        if self.inverse_goal:
+            self.data[self.y_binary_name] = self.base_data[self.y_name] >= self.y_pivot
+        else:
+            self.data[self.y_binary_name] = self.base_data[self.y_name] < self.y_pivot
+
+
 
     def _convert_cats(self) -> None:
         """ Converting categories to binary (one-hot encoding)
@@ -194,6 +233,7 @@ class BinaryDependenceModelData:
                 binary_name = col + '_' + str(val)
                 self.data[binary_name] = (self.base_data[col] == val).astype(int)
                 self.col_links[binary_name] = col
+                self.segment_sources[binary_name] = SegmentSource.PRIMARY_BINNING
         # self.data = self.data.drop(self.cat_cols, 1)
 
     def _convert_nums(self, bins: dict) -> None:
@@ -210,7 +250,14 @@ class BinaryDependenceModelData:
                     self.data[binary_name] = np.where(
                         (self.base_data[col] >= float(lb)) & (self.base_data[col] < float(rb)), 1, 0)
                 self.col_links[binary_name] = col
+                self.segment_sources[binary_name] = SegmentSource.PRIMARY_BINNING
         # self.data = self.data.drop(self.num_cols, 1)
+
+    def _make_segment_for_entire_dataset(self) -> None:
+        binary_name = ENTIRE_DATASET_SEGMENT_NAME
+        self.data[binary_name] = 1
+        self.col_links[binary_name] = ENTIRE_DATASET_SEGMENT_NAME
+        self.segment_sources[binary_name] = SegmentSource.ENTIRE_DATASET
 
     def convert_to_binary(self,
                           bins: Optional[dict] = None) -> None:
@@ -232,6 +279,7 @@ class BinaryDependenceModelData:
         self.bins = dict() if bins is None else bins
         self._convert_cats()
         self._convert_nums(self.bins)
+        self._make_segment_for_entire_dataset()
         self.is_data_converted = True
 
     def construct_partial_combs(self, selected_feature, consider_selected_base: bool = True):
@@ -280,9 +328,9 @@ class BinaryDependenceModelData:
                 binary_name = sel + '_AND_' + other
                 self.data[binary_name] = np.logical_and(self.data[sel], self.data[other]).astype(int)
                 self.col_links[binary_name] = sel if consider_selected_base else other
+                self.segment_sources[binary_name] = SegmentSource.PARTIAL_COMBINATION
 
-    # TODO: display progress in percentage instead of just combination levels
-    def construct_combs_up_to(self, comb_max_size: int) -> None:
+    def add_combinations_from_existing_segments(self, comb_max_size: int) -> None:
         """ Binary feature combinations of sizes up to comb_max_size are constructed as binary features.
                 These features equal 1 when all of its members equal 1.
 
@@ -315,20 +363,124 @@ class BinaryDependenceModelData:
 
         _comb_max_size = int(comb_max_size)
         if _comb_max_size < 2:
-            print("comb_max_size < 2, no features will be created")
+            logging.info("comb_max_size < 2, no features will be created")
             return
-        elif _comb_max_size > 5:
+        elif _comb_max_size > 2:
             logging.warning(f'Using high comb_max_size ({comb_max_size}), calculations might take some time.')
 
         unwanted = {self.y_name, self.y_binary_name}
-        binary_features = {c for c in list(self.data.columns) if c not in unwanted}
 
-        for comb_curr_size in range(2, _comb_max_size+1):
+        binary_features = set()
+        'c for c in list(self.data.columns) if c not in unwanted'
+        for c in list(self.data.columns):
+            if c in unwanted:
+                continue
+
+            # skip combination segments that were created by other methods or this one earlier calls
+            if self.segment_sources[c] != SegmentSource.PRIMARY_BINNING:
+                continue
+
+            binary_features.add(c)
+
+        new_segments = {}
+        for comb_curr_size in range(2, _comb_max_size + 1):
             logger.info(f'Working on combinations of level {comb_curr_size} of {_comb_max_size}')
             binary_combs = combinations(binary_features, comb_curr_size)
 
             for comb in binary_combs:
-                binary_name = '_AND_'.join(comb)
-                self.data[binary_name] = np.logical_and.reduce([self.data[col] for col in comb]).astype(int)
 
-                self.col_links[binary_name] = json.dumps(sorted(comb))
+                # if at least 2 of base segments have the same base column, it will make up an impossible segment
+                # (non-overlapping intervals)
+                base_columns = []
+                for binary_name in comb:
+                    base_columns.append(self.col_links[binary_name])
+                if len(set(base_columns)) != len(base_columns):
+                    continue
+
+                binary_name = '_AND_'.join(comb)
+                new_segments[binary_name] = np.logical_and.reduce([self.data[col] for col in comb]).astype(int)
+
+                # self.col_links[binary_name] = json.dumps(sorted(comb))
+
+                # getting segments' base_columns to construct a base_column for a combination
+                base_cols = []
+                for segment in comb:
+                    if segment in self.col_links:
+                        base_cols.append(self.col_links[segment])
+                    else:
+                        raise ValueError("Segment name not found in links")
+
+                # sorted so that parts from same 2 base columns always get a same combination
+                self.col_links[binary_name] = str(sorted(base_cols))
+                self.segment_sources[binary_name] = SegmentSource.COMBINATION_ON_EXISTING
+
+        # faster than inserting new column often (fragmented dataframe)
+        new_df = pd.DataFrame(new_segments)
+        new_df.index = self.data.index
+        self.data = self.data.join(new_df)
+
+    def add_combinations_from_decision_tree(
+            self,
+            comb_max_size: int | None,
+            comb_max_count: int
+    ) -> None:
+        res = get_optimized_combination_splits(
+            df=self.base_data.copy(),
+            cat_cols=list(self.cat_cols),
+            num_cols=list(self.num_cols),
+            target_name=self.y_name,
+
+            max_segments=comb_max_count,
+            max_depth=comb_max_size,
+        )
+
+        # Comparison will be by index, so it's helpful to check these weren't altered.
+        #  This still doesn't check that the order of values itself weren't altered
+        if not self.base_data.index.identical(self.data.index):
+            raise ValueError("Base data and binary data index mismatch")
+
+        for segment in res:
+            segment_raw_data = self.base_data.copy()
+            if segment['num']:
+                for num_feat in segment['num']:
+                    num_feat_name, num_feat_condition, num_feat_value = num_feat
+                    if num_feat_condition == '<=':
+                        segment_raw_data = segment_raw_data[segment_raw_data[num_feat_name] <= num_feat_value]
+                    elif num_feat_condition == '>':
+                        segment_raw_data = segment_raw_data[segment_raw_data[num_feat_name] > num_feat_value]
+                    else:
+                        raise ValueError("Unknown condition")
+            if segment['cat_include']:
+                cat_feat_name, allowed_cats = segment['cat_include']
+                segment_raw_data = segment_raw_data[segment_raw_data[cat_feat_name].isin(allowed_cats)]
+            if segment['cat_exclude']:
+                cat_feat_name, forbidden_cats = segment['cat_exclude']
+                segment_raw_data = segment_raw_data[~segment_raw_data[cat_feat_name].isin(forbidden_cats)]
+
+            segment_name = compile_splits_to_string_interval(segment)
+
+            self.data[segment_name] = self.base_data.index.isin(segment_raw_data.index)
+
+            # Same column name for all of these segments for now, don't have an idea how to split them yet
+            self.col_links[segment_name] = 'from_decision_tree_custom'
+            self.segment_sources[segment_name] = SegmentSource.DECISION_TREE_BINNING
+
+    def remove_segment_combinations(self):
+        to_exclude = []
+        for c in self.data.columns:
+            if c == self.y_name:
+                continue
+
+            if c == self.y_binary_name:
+                continue
+
+            if self.segment_sources[c] in (
+                SegmentSource.PARTIAL_COMBINATION,
+                SegmentSource.COMBINATION_ON_EXISTING,
+                SegmentSource.DECISION_TREE_BINNING
+            ):
+                continue
+
+        self.data.drop(to_exclude, axis=1)
+        for c in to_exclude:
+            del self.col_links[c]