+++ data_fast_insights/plotting/__init__.py
@@ -1,3 +1,4 @@
-from .essentials import plot_segments_dependence, plot_segments_central_tendency, plot_segments_basic_info
+from ._matplotlib_plots import plot_segments_dependence, plot_segments_central_tendency, plot_segments_basic_info
+from ._plotly_plots import create_feature_segments_plot_plotly
 
 __all__ = ['plot_segments_dependence', 'plot_segments_central_tendency', 'plot_segments_basic_info']