+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+### 0.4.5
+* Significantly speed up calculating dependence
+* Fix <ENTIER_DATASET> segment appearing in combinations
+
+### 0.4.4
+* Rename resulting df columns
+* Introduce sort argument for calculating results 
+
+### 0.4.3.
+* Allow inversed goals (experimental)
+
+### 0.4.2
+* Now sorting by target_delta_by_volume
+* Remove group_importance
+* Introduce decision tree based combination binning
+
+### 0.4.1
+* Rework base_cols of combined segments for better visual representation
+* Now sorting output dataframe by group_importance instead
+
+### 0.4
+* Switch to interactive plotly plotting
+* Misc refactoring and improvements
+
+### 0.3
+* Upgrade to work with newer versions of dependencies, except for scorecardpy, which was switched for optbinning
+* Various fixes and improvements
+
 ### 0.2.1 - 21 Apr 2021
 * Add group importance metric: see Calculated Metrics in [calculations description](data_fast_insights/doc/CALCULATIONS_DESCRIPTION.md)